add app icon and minor roomdb changes

+++ app/src/main/java/com/example/openrestaurant/model/CartItem.kt
@@ -3,5 +3,5 @@ package com.example.openrestaurant.model
 data class CartItem(
     var item: Item,
     var itemPath: String,
-    var quantity: Int = 0,
+    var quantity: Long = 0,
 )
